Summary:
Lifting RecordFunction up into the dispatcher code

Differential Revision: [D21374246](https://our.internmc.facebook.com/intern/diff/D21374246)

[ghstack-poisoned]

+++ torch/csrc/autograd/profiler.cpp
@@ -286,7 +286,7 @@ struct ProfilerThreadLocalState : public at::DebugInfoBase {
   }
 
   std::mutex state_mutex_;
-  std::unordered_map<uint16_t, std::shared_ptr<RangeEventList>>
+  std::unordered_map<uint64_t, std::shared_ptr<RangeEventList>>
       event_lists_map_;
   ProfilerConfig config_ = ProfilerConfig(ProfilerState::Disabled, false);
   at::CallbackHandle handle_;
